requirements.txt
Flask>=3.0.0
PyJWT>=2.8.0
Werkzeug>=3.0.0
requests>=2.31.0
pytest>=7.4.0
pandas>=2.0.0
openpyxl>=3.1.0
pytest-html>=4.1.0
pytest-cov>=4.1.0
flasgger>=0.9.5
allure-pytest>=2.13.0
jsonschema>=4.20.0
pydantic>=2.5.0
locust>=2.20.0
faker>=22.0.0
playwright>=1.40.0
pytest-playwright>=0.4.0

# Security Testing Dependencies (Fas 3)
bandit>=1.7.0
safety>=3.0.0
flask-talisman>=1.1.0
flask-limiter>=3.5.0